figurinhas/src/main/java/br/com/ada/figurinhas/service/impl/FigurinhaToSellServiceImpl.java
package br.com.ada.figurinhas.service.impl;

import br.com.ada.figurinhas.model.dto.FigurinhaToSellCreationDTO;
import br.com.ada.figurinhas.model.dto.FigurinhaToSellDTO;
import br.com.ada.figurinhas.model.dto.FigurinhaToSellUpdateDTO;
import br.com.ada.figurinhas.model.entity.Figurinha;
import br.com.ada.figurinhas.model.entity.FigurinhaToSell;
import br.com.ada.figurinhas.model.mapper.FigurinhaMapper;
import br.com.ada.figurinhas.model.mapper.FigurinhaToSellMapper;
import br.com.ada.figurinhas.repository.FigurinhaToSellRepository;
import br.com.ada.figurinhas.service.FigurinhaService;
import br.com.ada.figurinhas.service.FigurinhaToSellService;
import jakarta.persistence.EntityNotFoundException;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class FigurinhaToSellServiceImpl implements FigurinhaToSellService {

    private final FigurinhaToSellRepository repository;
    private final FigurinhaToSellMapper mapper;
    private final FigurinhaService figurinhaService;
    private final FigurinhaMapper figurinhaMapper;

    public FigurinhaToSellServiceImpl(final FigurinhaToSellRepository repository,
                                    final FigurinhaToSellMapper mapper,
                                    final FigurinhaService figurinhaService,
                                    final FigurinhaMapper figurinhaMapper) {
        this.repository = repository;
        this.mapper = mapper;
        this.figurinhaService = figurinhaService;
        this.figurinhaMapper = figurinhaMapper;
    }
    
    @Override
    public List<FigurinhaToSellDTO> findAll() {
        return mapper.parseListDTO(repository.findAll());
    }

    @Override
    public FigurinhaToSell findById(final String id) {
        Optional<FigurinhaToSell> optional = repository.findById(id);
        if (optional.isPresent()) {
            return optional.get();
        }
        throw new EntityNotFoundException();
    }

    @Override
    public FigurinhaToSellDTO create(final FigurinhaToSellCreationDTO creationDTO) {
        FigurinhaToSell entity = mapper.parseEntity(creationDTO);
        Figurinha figurinhaEntity = figurinhaMapper.parseEntity(
                figurinhaService.findById(creationDTO.getFigurinha().getId()));
        entity.setId(null);
        entity.setFigurinha(figurinhaEntity);
        entity = repository.save(entity);
        return mapper.parseDTO(entity);
    }

    @Override
    public FigurinhaToSellDTO edit(final String id, final FigurinhaToSellUpdateDTO updateDTO) {
        final Optional<FigurinhaToSell> optional = repository.findById(id);
        if (optional.isPresent()) {
            FigurinhaToSell entity = optional.get();
            entity.setId(id);
            entity.setPrice(updateDTO.getPrice());
            entity = repository.save(entity);
            return mapper.parseDTO(entity);
        }
        throw new EntityNotFoundException();
    }

    @Override
    public void delete(String id) {
        if (!repository.existsById(id)) {
            throw new EntityNotFoundException();
        }
        repository.deleteById(id);
    }

    @Override
    public void deleteByFigurinhaId(String figurinhaId) {
        Optional<FigurinhaToSell> optional = repository.findByFigurinhaId(figurinhaId);
        if (!optional.isPresent()) {
            throw new EntityNotFoundException();
        }
        repository.deleteById(optional.get().getId());
    }

    @Override
    public Optional<FigurinhaToSell> findByFigurinhaId(String figurinhaId) {
        return repository.findByFigurinhaId(figurinhaId);
    }
}
